+++ spec/dummy/config/initializers/money.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
 MoneyRails.configure do |config|
-  config.default_bank =
-    Money::Bank::VariableExchange.new(ActiveCurrency::RateStore.new)
+  config.default_bank = ActiveCurrency::Bank.new
 end